RJ-15: Refactoring StoryView to be used by different controllers.
